# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Cartes Scolaires",
        "description": "API complète pour la gestion des cartes scolaires",
        "contact": {
            "name": "Support API",
            "email": "support@cartes-scolaires.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Serveur de développement"
        },
        {
            "url": "https://api.cartes-scolaires.com/api/v1",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/api/etablissements": {
            "get": {
                "tags": [
                    "Etablissements"
                ],
                "summary": "Liste des établissements (Admin uniquement)",
                "operationId": "0a3bfcab07ae4e55e4e256999fb25850",
                "responses": {
                    "200": {
                        "description": "Liste des établissements"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/etablissements/{id}": {
            "get": {
                "tags": [
                    "Etablissements"
                ],
                "summary": "Détails d'un établissement",
                "operationId": "093e64c6cd789100f7c92a325864b67f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'établissement"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Etablissements"
                ],
                "summary": "Mettre à jour un établissement",
                "operationId": "b462adfc1bfe9dccccb88efb5dcaa973",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "adresse": {
                                        "type": "string"
                                    },
                                    "ville": {
                                        "type": "string"
                                    },
                                    "telephone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Établissement mis à jour"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/etablissements/{id}/logo": {
            "post": {
                "tags": [
                    "Etablissements"
                ],
                "summary": "Mettre à jour le logo",
                "operationId": "21180c75c638d150fc0587ee58b819a2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "logo": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logo mis à jour"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/etablissements/{id}/statistiques": {
            "get": {
                "tags": [
                    "Etablissements"
                ],
                "summary": "Statistiques détaillées d'un établissement",
                "operationId": "6518603811cc70cc0e5ba49b161717fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statistiques"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "name": "bearerAuth",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}